requirements.txt
numpy
pandas
matplotlib
seaborn
notebook
scikit-learn
tensorflow
keras
python-box==6.0.2
tqdm
ensure
joblib
types-PyYaML
scipy
Flask
Flask-Cors
dvc
-e .